backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database and storage
supabase==2.0.2
asyncpg==0.29.0

# Data processing and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# File handling and utilities
aiofiles==23.2.1
Pillow==10.1.0

# OMR Engine dependencies
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
scikit-image==0.22.0
pandas==2.1.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0